Do not use long type, it makes the code more platform dependent for no good reason.
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34195 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_pva.c
@@ -51,7 +51,7 @@
 
 typedef struct {
 	off_t offset;
-	long size;
+	int size;
 	uint8_t type;
 	uint8_t is_packet_start;
 	float pts;
@@ -62,7 +62,7 @@ typedef struct {
 	float last_audio_pts;
 	float last_video_pts;
 	float video_pts_after_prebytes;
-	long video_size_after_prebytes;
+	int video_size_after_prebytes;
 	uint8_t prebytes_delivered;
 	uint8_t just_synced;
 	uint8_t synced_stream_id;
